Pruning some abandoned ones (#28)
IRCAnywhere - Abandoned in 2015
Tapchat - Abandoned in 2014
nirc - Abandoned in 2015
aIRChat - Abandoned in 2015
Jerk - Abandoned in 2012
CamperVan - Last updated in 2014. In 2015 Basecamp 3 was released and Campfire was retired as a separate product and fused into Basecamp. Basecamp 3 API is not compatible with the older APIs: https://github.com/basecamp/bc3-api Thus, I doubt CamperVan has worked in a long time.